scheduler: add pending_cnt for task to make pending scheduling possible
When one instance of a task is queued(queued, pending, running, or
preempted), it is possible that another instance of this same task
scheduling reqire will come before the previous one is completed. In
today's design, this second instance may be dropped(e.g. for
pipeline_copy() task) and unexpected error will happen.

Here, we add a pending_cnt for the task, accumulate it when needed, and
schedule it when the previous one is finished.

Signed-off-by: Keyon Jie <yang.jie@linux.intel.com>